math-engine/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ─── Validation ───────────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
jsonschema==4.21.1

# ─── Math & Science ───────────────────────────────────────────────────────────
sympy==1.14
numpy==1.26.4
matplotlib==3.10.1
networkx==2.8.8
scipy
mpmath==1.3.0
pyformlang==1.0.7
pandas
matplotlib-venn

# ─── AI & APIs ────────────────────────────────────────────────────────────────
openai
requests

# ─── Image / OCR ──────────────────────────────────────────────────────────────
easyocr
pillow
opencv-python-headless

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest
httpx

# ─── Linting ──────────────────────────────────────────────────────────────────
flake8
